+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file, and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* Added customizable timeout to notebook execution
+* Changed the notebook execution template to use `pybryt.tracing_on` and `pybryt.tracing_off`
+
 ## 0.1.7 - 2021-06-28
 
 * Fixed `pybryt execute` output per [#94](https://github.com/microsoft/pybryt/issues/94)